updates to Sendgrid setup to include new email address
