+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "shugas",
-  "version": "0.1.0",
+  "version": "1.02.03",
   "main": "index.js",
   "scripts": {
     "cdk": "cdk",
